CHART-4759 Limits caller ID dropdown for non admin users
This implements a naming convention that feels a somewhat brittle.

It would not function as expected if multiple users with common first
and last names were present in the system. For our use cases it should
be fine.

In the future it would be nice to add an "assignee" field to a flow so
that we could just look up a flow by an assignee user ID. But for now,
this adequately filters down.